# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Recommendation Engine API
  description: |
    A comprehensive recommendation engine API that provides personalized content recommendations
    using multiple algorithms including semantic search, collaborative filtering, and graph-based approaches.
  version: 1.0.0
  contact:
    name: API Support
    email: api-support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.recommendations.example.com/api/v1
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /users/{userId}/interactions:
    get:
      summary: Get user interaction history
      description: Retrieve interaction history for a specific user
      operationId: getUserInteractions
      tags:
        - User Interactions
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: type
          in: query
          schema:
            type: string
            enum: [rating, like, dislike, share, click, view, search, browse]
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: User interactions retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  interactions:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserInteraction'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
  schemas:
...
    UserInteraction:
      type: object
      properties:
        userId:
          type: string
          format: uuid
          description: ID of the user performing the interaction
        itemId:
          type: string
          description: ID of the content item being interacted with
        interactionType:
          type: string
          enum: [rating, like, dislike, share, click, view, search, browse]
          description: Type of interaction
        value:
          type: number
          description: Numeric value associated with interaction (rating score, duration, etc.)
          minimum: 0
        timestamp:
          type: string
          format: date-time
          description: When the interaction occurred
        sessionId:
          type: string
          format: uuid
          description: Session ID for grouping related interactions
        context:
          type: object
          additionalProperties: true
          description: Additional context about the interaction
          example:
            source: "search_results"
            device: "mobile"
            duration: 45.5
      required:
        - userId
        - itemId
        - interactionType
        - timestamp
...
    Pagination:
      type: object
      properties:
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer
        hasMore:
          type: boolean
...